Add refactored object_data and stat_tracker
